FEAT: -> Correção de metodo CalculateCriterionGradeByGroup

+++ hackweek-backend/dtos/GroupRatingDto.cs
@@ -5,7 +5,7 @@ namespace hackweek_backend.DTOs
     public class GroupRatingDto
     {
         public int Id { get; set; }
-        public uint Grade { get; set; }
+        public double Grade { get; set; }
         public int GroupId { get; set; }
         public int PropositionCriterionId { get; set; }
